refactor(tui): colocate Screen::Gists into tui/screens/gists.rs
Sixth screen migration for #287 (Phase 2). Moved into the new
tui/screens/gists.rs: handle_key_gists, gists_guard (the #288 shared
guard, Gists-exclusive), build_gists_vm, render_gists_vm,
gists_palette_items.

gist_group_row_label (render.rs) is shared by build_gists_vm and mod.rs's
gists_hscroll_max (a generic AppState hscroll helper, not tied to any one
screen module), so it stays in render.rs. apply_filter_edit/FilterKey
(already pub(crate) from the Pins migration) are reused here too — Gists
has the same inline-filter pattern as Pins.

star_toggle_intent (keys.rs) widened from private to pub(crate): shared by
List/Detail (still in keys.rs) and now Gists (moved) — no change in
ownership, just visibility to permit the cross-module call.

Pure move — bodies and existing tests unchanged.

Part of #287
