+++ src/tui/screens/mod.rs
@@ -6,6 +6,7 @@
 
 pub(crate) mod config;
 pub(crate) mod diff;
+pub(crate) mod gists;
 pub(crate) mod help;
 pub(crate) mod pins;
 pub(crate) mod preview;
